+++ packages/hummingbird/node/hooks/createPages/query.js
@@ -4,6 +4,13 @@ fragment NodeFields on Wings_Node {
   title
   resourceType
   slug
+  featured {
+    title
+    description
+    image {
+      url
+    }
+  }
   locale {
     id
     name
@@ -40,23 +47,33 @@ fragment NodeFields on Wings_Node {
   status
   nodeType
   platforms {
-    all {
+    search {
       title
       description
-      imageUrl
     }
     facebook {
       title
       description
-      imageUrl
+      image {
+        url
+      }
     }
     twitter {
       title
       description
-      imageUrl
+      image {
+        url
+      }
     }
     whatsapp {
       text
+    }
+    meta {
+      tag
+      attributes {
+        key
+        value
+      }
     }
   }
 }
@@ -65,11 +82,21 @@ fragment CampaignFields on Wings_Campaign {
   intro
   description
   submissionSchema
+  settings {
+    legal {
+      terms {
+        url
+      }
+      privacyPolicy {
+        url
+      }
+    }
+  }
 }
 
 {
   wings {
-    articles: entries(type: "article") {
+    articles: entries(selector: { typeId: { eq: "article" } }, first: 0) {
       edges {
         node {
           ...NodeFields
@@ -80,7 +107,7 @@ fragment CampaignFields on Wings_Campaign {
         }
       }
     }
-    pages: entries(type: "page") {
+    pages: entries(selector: { typeId: { eq: "page" } }, first: 0) {
       edges {
         node {
           ...NodeFields
@@ -91,7 +118,7 @@ fragment CampaignFields on Wings_Campaign {
         }
       }
     }
-    events {
+    events(first: 0) {
       edges {
         node {
           ...NodeFields
@@ -115,7 +142,15 @@ fragment CampaignFields on Wings_Campaign {
         }
       }
     }
-    petitions {
+    signups(first: 0) {
+      edges {
+        node {
+          ...NodeFields
+          ...CampaignFields
+        }
+      }
+    }
+    petitions(first: 0) {
       edges {
         node {
           ...NodeFields
@@ -125,7 +160,7 @@ fragment CampaignFields on Wings_Campaign {
         }
       }
     }
-    fundraisers {
+    fundraisers(first: 0) {
       edges {
         node {
           ...NodeFields